event updated

+++ zzio/src/main/java/com/game/zzio/domain/CreateRankRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CreateRankRequest {
     private String nickName;
-    private Integer score;
+    private Long score;
 }
